feat: add Sqflite-backed SyncStorageAdapter for relay_sync
- Introduced SqfliteSyncStorage for persistent task storage using Sqflite.
- Added README.md with setup instructions, schema details, and migration examples.
- Implemented default database factory for Sqflite and a stub for non-Flutter environments.
- Developed comprehensive methods for task management including save, update, delete, and retrieval.
- Established schema creation with necessary indexes for efficient querying.
- Included streaming capabilities for task updates and status counts.

+++ packages/relay_sync/lib/src/retry/exponential_backoff_retry_policy.dart
@@ -61,7 +61,7 @@ class ExponentialBackoffRetryPolicy extends RetryPolicy {
     }
 
     final exponential = initialDelay.inMicroseconds * pow(multiplier, retryCount);
-    final cappedMicros = min(exponential, maxDelay.inMicroseconds.toDouble());
+    final cappedMicros = min(exponential, maxDelay.inMicroseconds.toDouble()).toDouble();
     final adjustedMicros = _applyJitter(cappedMicros);
     return now.add(Duration(microseconds: adjustedMicros.round()));
   }
